+++ src/PublishPSResource.cs
@@ -294,11 +294,11 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         }
         private string _nuspec;
 
-
-        private Hashtable dependencies = new Hashtable();
         private NuGetVersion pkgVersion = null;
         private bool isScript;
         private string pkgName;
+
+        private static char[] PathSeparators = new [] { System.IO.Path.DirectorySeparatorChar, System.IO.Path.AltDirectorySeparatorChar };
 
         protected override void ProcessRecord()
         {
@@ -337,6 +337,7 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
             }
 
             // if user does not specify that they want to use a nuspec they've created, we'll create a nuspec
+            var dependencies = new Hashtable();
             if (string.IsNullOrEmpty(_nuspec))
             {
                 _nuspec = createNuspec(outputDir, moduleFileInfo);
@@ -370,6 +371,15 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
             // find repository
             var r = new RespositorySettings();
             var repositoryUrl = r.Read(new[] { _repository });
+
+            if (!repositoryUrl.Any())
+            {
+                var message = String.Format("The resource repository '{0}' is not a registered. Please run 'Register-PSResourceRepository' in order to publish to this repository.", _repository);
+                var ex = new ArgumentException(message);
+                var repositoryNotFound = new ErrorRecord(ex, "repositoryNotFound", ErrorCategory.ObjectNotFound, null);
+
+                this.ThrowTerminatingError(repositoryNotFound);
+            }
 
             if (!_skipDependenciesCheck)
             {
@@ -407,12 +417,14 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
                 // Create subdirectory structure in temp folder
                 foreach (string dir in System.IO.Directory.GetDirectories(_path, "*", System.IO.SearchOption.AllDirectories))
                 {
-                    System.IO.Directory.CreateDirectory(System.IO.Path.Combine(outputDir, dir.Substring(_path.Length + 1)));
+                    var dirName = dir.Substring(_path.Length).Trim(PathSeparators);
+                    System.IO.Directory.CreateDirectory(System.IO.Path.Combine(outputDir, dirName));
                 }
                 // Copy files over to temp folder
-                foreach (string file_name in System.IO.Directory.GetFiles(_path, "*", System.IO.SearchOption.AllDirectories))
+                foreach (string fileNamePath in System.IO.Directory.GetFiles(_path, "*", System.IO.SearchOption.AllDirectories))
                 {
-                    System.IO.File.Copy(file_name, System.IO.Path.Combine(outputDir, file_name.Substring(_path.Length + 1)));
+                    var fileName = fileNamePath.Substring(_path.Length).Trim(PathSeparators);
+                    System.IO.File.Copy(fileNamePath, System.IO.Path.Combine(outputDir, fileName));
                 }
             }
 
@@ -613,28 +625,24 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
                 metadataElement.AppendChild(element);
             }
 
-            if (parsedMetadataHash.ContainsKey("requiredmodules"))
+            var requiredModules = ParseRequiredModules(parsedMetadataHash);
+            if (requiredModules != null)
             {
-                dependencies = (Hashtable)parsedMetadataHash["requiredmodules"];
+                XmlElement dependenciesElement = doc.CreateElement("dependencies", nameSpaceUri);
 
-                if (dependencies != null)
+                foreach (Hashtable dependency in requiredModules)
                 {
-                    XmlElement dependenciesElement = doc.CreateElement("dependencies", nameSpaceUri);
+                    XmlElement element = doc.CreateElement("dependency", nameSpaceUri);
 
-                    foreach (Hashtable dependency in dependencies)
+                    element.SetAttribute("id", dependency["ModuleName"].ToString());
+                    if (!string.IsNullOrEmpty(dependency["ModuleVersion"].ToString()))
                     {
-                        XmlElement element = doc.CreateElement("dependency", nameSpaceUri);
-
-                        element.SetAttribute("id", dependency["ModuleName"].ToString());
-                        if (!string.IsNullOrEmpty(dependency["ModuleVersion"].ToString()))
-                        {
-                            element.SetAttribute("version", dependency["ModuleVersion"].ToString());
-                        }
-
-                        dependenciesElement.AppendChild(element);
+                        element.SetAttribute("version", dependency["ModuleVersion"].ToString());
                     }
-                    metadataElement.AppendChild(dependenciesElement);
+
+                    dependenciesElement.AppendChild(element);
                 }
+                metadataElement.AppendChild(dependenciesElement);
             }
             
             packageElement.AppendChild(metadataElement);
@@ -648,6 +656,43 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
             return nuspecFullName;
         }
 
+        private Hashtable[] ParseRequiredModules(Hashtable parsedMetadataHash)
+        {
+            if (!parsedMetadataHash.ContainsKey("requiredmodules"))
+            {
+                return null;
+            }
+
+            var requiredModules = parsedMetadataHash["requiredmodules"];
+        
+            // Required modules can be:
+            //  a. An array of hash tables of module name and version
+            //  b. A single hash table of module name and version
+            //  c. A string array of module names
+            //  d. A single string module name
+
+            if (LanguagePrimitives.TryConvertTo<Hashtable[]>(requiredModules, out Hashtable[] moduleList))
+            {
+                return moduleList;
+            }
+
+            if (LanguagePrimitives.TryConvertTo<string[]>(requiredModules, out string[] moduleNames))
+            {
+                var listHashtable = new List<Hashtable>();
+                foreach (var modName in moduleNames)
+                {
+                    listHashtable.Add(
+                        new Hashtable() {
+                            { "ModuleName", modName },
+                            { "ModuleVersion", string.Empty }
+                        });
+                }
+
+                return listHashtable.ToArray();
+            }
+
+            return null;
+        }
 
         private void ParseScriptMetadata(Hashtable parsedMetadataHash, FileInfo moduleFileInfo)
         {