+++ src/RegisterPSResourceRepository.cs
@@ -70,20 +70,6 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         private Uri _url;
 
         /// <summary>
-        /// Specifies a user account that has rights to find a resource from a specific repository.
-        /// </summary>
-        [Parameter(ValueFromPipelineByPropertyName = true, ParameterSetName = "NameParameterSet")]
-        public PSCredential Credential
-        {
-            get
-            { return _credential; }
-
-            set
-            { _credential = value; }
-        }
-        private PSCredential _credential;
-
-        /// <summary>
         /// Registers the PowerShell Gallery.
         /// </summary>
         [Parameter(Mandatory = true, ParameterSetName = "PSGalleryParameterSet")]